split into ability specific and global
