+++ src/MacroTools/Hazards/RecurrentSpellHazard.cs
@@ -1,4 +1,5 @@
-﻿using MacroTools.SpellSystem;
+﻿using MacroTools.DummyCasters;
+using MacroTools.SpellSystem;
 using static War3Api.Common;
 
 namespace MacroTools.Hazards
