+++ src/MacroTools/MacroTools.csproj
@@ -18,4 +18,5 @@
         <PackageReference Include="WCSharp.Shared" Version="2.2.0" />
     </ItemGroup>
 
+
 </Project>
