+++ src/MacroTools/Spells/ChannelAnySpell.cs
@@ -1,4 +1,5 @@
-﻿using MacroTools.SpellSystem;
+﻿using MacroTools.DummyCasters;
+using MacroTools.SpellSystem;
 using WCSharp.Shared.Data;
 using static War3Api.Common;
