+++ src/MacroTools/Spells/MultiTargetRecall.cs
@@ -1,4 +1,5 @@
 ﻿using System.Linq;
+using MacroTools.DummyCasters;
 using MacroTools.Extensions;
 using MacroTools.SpellSystem;
 using WCSharp.Shared.Data;
