+++ src/MacroTools/Spells/Stomp.cs
@@ -1,4 +1,5 @@
-﻿using MacroTools.Extensions;
+﻿using MacroTools.DummyCasters;
+using MacroTools.Extensions;
 using MacroTools.SpellSystem;
 using WCSharp.Effects;
 using WCSharp.Shared.Data;
