+++ src/WarcraftLegacies.Source/Setup/Spells/DalaranSpellSetup.cs
@@ -1,4 +1,5 @@
-﻿using MacroTools.PassiveAbilitySystem;
+﻿using MacroTools.DummyCasters;
+using MacroTools.PassiveAbilitySystem;
 using MacroTools.Spells;
 using MacroTools.SpellSystem;
 using WarcraftLegacies.Source.Setup.FactionSetup;
