+++ src/WarcraftLegacies.Source/Setup/Spells/ScourgeSpellSetup.cs
@@ -1,4 +1,5 @@
-﻿using MacroTools.PassiveAbilities;
+﻿using MacroTools.DummyCasters;
+using MacroTools.PassiveAbilities;
 using MacroTools.PassiveAbilitySystem;
 using MacroTools.Spells;
 using MacroTools.SpellSystem;
